move flagOn to authUtils
