+++ src/auth/token.ts
@@ -68,15 +68,6 @@ export async function readAuthData(): Promise<
 }
 
 /**
- * Returns true if the specified flag is on for the current user.
- * @param flag the feature flag to check
- */
-export async function flagOn(flag: string): Promise<boolean> {
-  const authData = await readAuthData();
-  return authData.flags?.includes(flag);
-}
-
-/**
  * Return the native PixieBrix API token (issued by the PixieBrix API).
  */
 export async function getExtensionToken(): Promise<string | undefined> {
