+++ src/services/errorService.ts
@@ -23,7 +23,7 @@ import {
 } from "@/errors/errorHelpers";
 import { allowsTrack } from "@/telemetry/dnt";
 import { uuidv4, validateSemVerString } from "@/types/helpers";
-import { flagOn, getUserData } from "@/auth/token";
+import { getUserData } from "@/auth/token";
 import {
   isAppRequest,
   selectAbsoluteUrl,
@@ -37,6 +37,7 @@ import { type SemVerString } from "@/types/registryTypes";
 import { type MessageContext } from "@/types/loggerTypes";
 import { isObject } from "@/utils/objectUtils";
 import type { UnknownObject } from "@/types/objectTypes";
+import { flagOn } from "@/auth/authUtils";
 
 const EVENT_BUFFER_DEBOUNCE_MS = 2000;
 const EVENT_BUFFER_MAX_MS = 10_000;
